Add rudimentary support for working with the Syncthing API
This lets us ask Syncthing for which directories we should search for conflicts
in. This also alters the `emacs-conflict-resolve-conflicts` call to make the
directory optional. If not provided, all directories owned by Syncthing are
scanned.

Adds Syncthing conflict cache. Since searching large (i.e. multi-gigabyte)
directories can take a while, the results are cached for a speedier workflow. A
`emacs-conflict-syncthing-conflicts-refresh-cache` command is provided to
refresh the cache. When a conflict is resolved its entry in the cache is
cleared.

Adds a `emacs-conflict-remove-conflict` that lets the user interactively clear
conflicts if they're invalid or otherwise impractical to handle via this
package. Large binary files can be rejected by `ediff` and is cumbersome to work
with in Emacs anyways.

Simple management of the Syncthing API key is added - it's prompted for if it is
not set. Also, since it's very plausible that a user that would be interested in
this package is also syncing their dotfiles already, there's support for
multiple API keys (keyed on the result of `(system-name)`). This was discovered
to be a problem during development of these features, heh.

Finally, add a keymap with the interactive features bound. A suggested bind for
it could be `C-x c`, with the _conflict_ mnemonic.